requirements.txt
beautifulsoup4
ebird-api
requests
PyGithub